+++ spython/oci/cmd/__init__.py
@@ -13,7 +13,6 @@ def generate_oci_commands():
     from spython.oci import OciImage
 
     from spython.main.base.logger import println
-    from spython.utils import run_command as run_cmd
 
     # run_command uses run_cmd, but wraps to catch error
     from spython.main.base.command import ( run_command, send_command )
@@ -21,7 +20,7 @@ def generate_oci_commands():
 
     # Oci Command Groups
     from .mounts import ( mount, umount )
-    from .states import ( kill, state, start, resume )
+    from .states import ( kill, state, start, pause, resume, _state_command  )
     from .actions import ( attach, create, delete, execute, run, _run, update )
 
     # Oci Commands
@@ -34,15 +33,16 @@ def generate_oci_commands():
     OciImage.create = create
     OciImage.delete = delete
     OciImage.execute = execute
+    OciImage.update = update
     OciImage.kill = kill
     OciImage.run = run
     OciImage._run = _run
-    OciImage.update = update
+    OciImage._state_command = _state_command
 
     OciImage.RobotNamer = RobotNamer()
-    OciImage.run_command = run_cmd
     OciImage._send_command = send_command # send and disregard stderr, stdout
     OciImage._run_command = run_command
     OciImage._println = println
- 
+    OciImage.OciImage = OciImage
+
     return OciImage