+++ spython/oci/cmd/states.py
@@ -10,7 +10,7 @@ from spython.logger import bot
 import json
 
 
-def state(self, container_id=None, sudo=False, sync_socket=False):
+def state(self, container_id=None, sudo=None, sync_socket=None):
 
     ''' get the state of an OciImage, if it exists. The optional states that
         can be returned are created, running, stopped or (not existing).
@@ -30,6 +30,7 @@ def state(self, container_id=None, sudo=False, sync_socket=False):
         state: a parsed json of the container state, if exists. If the
                container is not found, None is returned.
     '''
+    sudo = self._get_sudo(sudo)
     container_id = self.get_container_id(container_id)
 
     # singularity oci state
@@ -41,15 +42,49 @@ def state(self, container_id=None, sudo=False, sync_socket=False):
     # Finally, add the container_id
     cmd.append(container_id)
 
-    # Generate the instance
-    result = self._run_command(cmd, sudo=sudo, quiet=True)
+    # Get the instance state
+    result = self._run_command(cmd, sudo=sudo, quiet=True, return_result=True)
 
-    # Return the state object to the user
-    if result['return_code'] == 0:
-        return json.loads(result['message'][0])
-    
+    # If successful, a string is returned to parse
+    if isinstance(result, str):
+        return json.loads(result)
 
-def start(self, container_id=None, sudo=False):
+
+def _state_command(self, container_id=None, command='start', sudo=None):
+
+    ''' A generic state command to wrap pause, resume, kill, etc., where the
+        only difference is the command. This function will be unwrapped if the
+        child functions get more complicated (with additional arguments).
+
+        Equivalent command line example:      
+           singularity oci <command> <container_ID>
+           
+        Parameters
+        ==========
+        container_id: the id to start.
+        command: one of start, resume, pause, kill, defaults to start.
+        sudo: Add sudo to the command. If the container was created by root,
+              you need sudo to interact and get its state.
+
+        Returns
+        =======
+        return_code: the return code to indicate if the container was started.
+    '''
+    sudo = self._get_sudo(sudo)
+    container_id = self.get_container_id(container_id)
+
+    # singularity oci state
+    cmd = self._init_command(command)
+
+    # Finally, add the container_id
+    cmd.append(container_id)
+
+    # Run the command, return return code
+    return self._run_and_return(cmd, sudo)
+
+
+
+def start(self, container_id=None, sudo=None):
 
     ''' start a previously invoked OciImage, if it exists.
 
@@ -66,19 +101,10 @@ def start(self, container_id=None, sudo=False):
         =======
         return_code: the return code to indicate if the container was started.
     '''
-    container_id = self.get_container_id(container_id)
-
-    # singularity oci state
-    cmd = self._init_command('start')
-
-    # Finally, add the container_id
-    cmd.append(container_id)
-
-    # Run the command, return return code
-    return self._run_and_return(cmd, sudo)
+    return self._state_command(container_id, sudo=sudo)
 
 
-def kill(self, container_id=None, sudo=False):
+def kill(self, container_id=None, sudo=None):
 
     ''' stop (kill) a started OciImage container, if it exists
 
@@ -95,19 +121,10 @@ def kill(self, container_id=None, sudo=False):
         =======
         return_code: the return code to indicate if the container was killed.
     '''
-    container_id = self.get_container_id(container_id)
-
-    # singularity oci state
-    cmd = self._init_command('kill')
-
-    # Finally, add the container_id
-    cmd.append(container_id)
-
-    # Run the command, return return code
-    return self._run_and_return(cmd, sudo)
+    return self._state_command(container_id, command='kill', sudo=sudo)
 
 
-def resume(self, container_id=None, sudo=False):
+def resume(self, container_id=None, sudo=None):
     ''' resume a stopped OciImage container, if it exists
 
         Equivalent command line example:      
@@ -123,13 +140,23 @@ def resume(self, container_id=None, sudo=False):
         =======
         return_code: the return code to indicate if the container was resumed.
     '''
-    container_id = self.get_container_id(container_id)
+    return self._state_command(container_id, command='resume', sudo=sudo)
 
-    # singularity oci state
-    cmd = self._init_command('resume')
 
-    # Finally, add the container_id
-    cmd.append(container_id)
+def pause(self, container_id=None, sudo=None):
+    ''' pause a running OciImage container, if it exists
 
-    # Run the command, return return code
-    return self._run_and_return(cmd, sudo)
+        Equivalent command line example:      
+           singularity oci pause <container_ID>
+           
+        Parameters
+        ==========
+        container_id: the id to stop.
+        sudo: Add sudo to the command. If the container was created by root,
+              you need sudo to interact and get its state.
+
+        Returns
+        =======
+        return_code: the return code to indicate if the container was paused.
+    '''
+    return self._state_command(container_id, command='pause', sudo=sudo)