+++ spython/utils/__init__.py
@@ -10,7 +10,6 @@ from .terminal import (
     check_install, 
     get_installdir,
     get_singularity_version,
-    send_command,
     stream_command,
     run_command,
     format_container_name,